Meas::Value now double.

+++ libmemseries/meas.h
@@ -13,7 +13,7 @@ namespace memseries {
 	typedef uint64_t Time;
 	typedef uint64_t Id;
 	typedef uint64_t Flag;
-	typedef uint64_t Value;
+	typedef double Value;
 	typedef std::vector<Id> IdArray;
 	typedef std::set<Id> IdSet;
